Move the various local scripts into `src-script` directory.
