+++ src-script/license-add.js
