+++ src-script/license-whitelist.txt
