+++ src-script/sdk-regen.js
@@ -4,16 +4,16 @@ import {
   closeDatabase,
   initDatabase,
   loadSchema,
-} from './src-electron/db/db-api.js'
+} from '../src-electron/db/db-api.js'
 import {
   schemaFile,
   sqliteTestFile,
   appDirectory,
   setDevelopmentEnv,
-} from './src-electron/util/env.js'
-import { runSdkGeneration } from './src-electron/sdk-gen/sdk-gen.js'
-import { loadZcl } from './src-electron/zcl/zcl-loader.js'
-import { zclPropertiesFile } from './src-electron/main-process/args.js'
+} from '../src-electron/util/env.js'
+import { runSdkGeneration } from '../src-electron/sdk-gen/sdk-gen.js'
+import { loadZcl } from '../src-electron/zcl/zcl-loader.js'
+import { zclPropertiesFile } from '../src-electron/main-process/args.js'
 
 setDevelopmentEnv()
 var file = sqliteTestFile(42)
